Implement Display trait on ScaleValue.

+++ shared/src/types.rs
@@ -2,6 +2,6 @@ pub type Volume = f32;
 pub type Freq = f32;
 pub type Semitones = f32;
 pub type Beats = f32;
-pub type Octave = u8;
+pub type Octave = u32;
 pub type PitchValue = u32;
